ast 모듈을 좀 더 작은 분류의 모듈들로 분리
