+++ src/ast/expr.ts
@@ -1,0 +1,395 @@
+// NOTE: 이 파일에서 String, Number, Boolean은 javascript builtin object가 아닙니다.
+import { AstListMixin, AstNode, AstNodeList, astList, child } from './base';
+
+// expression
+export class Expressions extends AstNodeList<Expression> {
+    get repr() {
+        return this.childNodes.map(childNode => childNode.repr).join(' ');
+    }
+}
+export class Expression extends AstNode {
+    type: any;
+    get isConstant() { return false; }
+    fold(): Expression { return this; }
+    get repr(): string {
+        throw new Error('unimplemented');
+    }
+}
+Expression.prototype.type = null;
+
+@astList('args')
+export class CallInfo extends AstNode {
+    def: AstNode;
+    args: Expression[];
+    constructor(def: AstNode) {
+        super();
+        this.def = def;
+        this.args = [];
+    }
+}
+export interface CallInfo extends AstListMixin<Expression> {}
+
+export interface CallLike extends Expression {
+    expressions: Expressions;
+    callInfo: CallInfo | null;
+}
+
+export interface ModuleCallLike extends CallLike {
+    target: Name;
+}
+
+export class Call extends Expression implements CallLike {
+    @child expressions: Expressions;
+    @child callInfo: CallInfo | null;
+
+    constructor(expressions: Expressions) {
+        super();
+        this.expressions = expressions; // analyzer 패스를 거친 뒤로는 무의미
+        this.callInfo = null; // analyzer 패스를 거친 뒤부터 접근 가능
+    }
+    fold() {
+        for (let arg of this.callInfo?.args || []) {
+            arg.fold();
+        }
+        return this; // TODO: fold call
+    }
+    get repr() {
+        return this.expressions?.repr ?? '';
+    }
+}
+
+export class ModuleCall extends Expression implements ModuleCallLike {
+    @child target: Name;
+    @child expressions: Expressions;
+    @child callInfo: CallInfo | null;
+
+    constructor(target: Name, expressions: Expressions) {
+        super();
+        this.target = target;
+        this.expressions = expressions; // analyzer 패스를 거친 뒤로는 무의미
+        this.callInfo = null; // analyzer 패스를 거친 뒤부터 접근 가능
+    }
+    fold() {
+        for (let arg of this.callInfo?.args || []) {
+            arg.fold();
+        }
+        return this; // TODO: fold call
+    }
+    get repr() {
+        return `@${ this.target.repr } ${ this.expressions?.repr }`;
+    }
+}
+
+export class CallBind extends Expression implements CallLike {
+    @child expressions: Expressions;
+    @child callInfo: CallInfo | null;
+
+    constructor(expressions: Expressions) {
+        super();
+        this.expressions = expressions;
+        this.callInfo = null;
+    }
+    get repr() {
+        return `결속 ${ this.expressions?.repr }`;
+    }
+}
+
+export class ModuleCallBind extends Expression implements ModuleCallLike {
+    @child target: any;
+    @child expressions: Expressions;
+    @child callInfo: CallInfo | null;
+
+    constructor(target: any, expressions: Expressions) {
+        super();
+        this.target = target;
+        this.expressions = expressions;
+        this.callInfo = null;
+    }
+    get repr() {
+        return `결속 @${ this.target.repr } ${ this.expressions?.repr }`;
+    }
+}
+export class Question extends Expression {
+    get repr() {
+        return '?';
+    }
+}
+
+// primitive
+export class Primitive extends Expression {
+    value: any;
+    constructor(value: any) { super(); this.value = value; }
+    get isConstant() { return true; }
+    clone() { return new (this.constructor as any)(this.value); }
+    get repr() {
+        return `${ this.value }`;
+    }
+}
+export class Name extends Primitive {
+    call: boolean;
+    constructor(value: any) {
+        super(value);
+        this.call = false; // true: 식별자 하나짜리 약속일지도 모름
+                           // analyzer 패스에서 사용되고 난 이후에는 의미가 없다.
+    }
+    get isConstant() {
+        // TODO: 상수 접기 구현
+        return false;
+    }
+}
+export class String extends Primitive {
+    get repr() {
+        return `"${ this.value }"`;
+    }
+} String.prototype.type = String;
+
+export abstract class Number extends Primitive {}
+
+export class Integer extends Number {} Integer.prototype.type = Integer;
+export class Float extends Number {} Float.prototype.type = Float;
+
+export class Boolean extends Primitive {} Boolean.prototype.type = Boolean;
+export class Void extends Primitive {
+    get repr() {
+        return '()';
+    }
+} Void.prototype.type = Void;
+
+// etc
+export class Range extends Expression {
+    @child start: Expression;
+    @child stop: Expression;
+    constructor(start: Expression, stop: Expression) {
+        super();
+        this.start = start;
+        this.stop = stop;
+    }
+}
+Range.prototype.type = Range;
+
+@astList('items')
+export class List extends Expression {
+    items: Expression[];
+    constructor() {
+        super();
+        this.items = [];
+    }
+}
+List.prototype.type = List;
+export interface List extends AstListMixin<Expression> {}
+
+@astList('items')
+export class Dict extends Expression {
+    items: DictKeyValue[];
+    constructor() {
+        super();
+        this.items = [];
+    }
+}
+Dict.prototype.type = Dict;
+export interface Dict extends AstListMixin<DictKeyValue> {}
+
+export class DictKeyValue extends AstNode {
+    @child key: Name;
+    @child value: Expression;
+    constructor(key: Name, value: Expression) {
+        super();
+        this.key = key;
+        this.value = value;
+    }
+}
+
+// unary operator
+export abstract class UnaryOperator extends Expression {
+    @child rhs: Expression;
+    constructor(rhs: Expression) {
+        super();
+        this.rhs = rhs;
+    }
+    get isConstant() { return this.rhs.isConstant; }
+}
+// arithmetical
+export class UnaryPlus extends UnaryOperator {
+    fold() {
+        if (this.isConstant) {
+            let rhs = this.rhs.fold();
+            if (rhs instanceof Integer)
+                return this.replace(new Integer(rhs.value));
+            if (rhs instanceof Number)
+                return this.replace(new Float(rhs.value));
+        }
+        return this;
+    }
+}
+export class UnaryMinus extends UnaryOperator {
+    fold() {
+        if (this.isConstant) {
+            let rhs = this.rhs.fold();
+            if (rhs instanceof Integer)
+                return this.replace(new Integer(-rhs.value));
+            if (rhs instanceof Number)
+                return this.replace(new Float(-rhs.value));
+        }
+        return this;
+    }
+}
+
+// binary opeartor
+export abstract class BinaryOperator extends Expression {
+    @child lhs: Expression;
+    @child rhs: Expression;
+    constructor(lhs: Expression, rhs: Expression) {
+        super();
+        this.lhs = lhs;
+        this.rhs = rhs;
+    }
+    get isConstant() { return this.lhs.isConstant && this.rhs.isConstant; }
+}
+export class Access extends BinaryOperator {}
+export class DotAccess extends BinaryOperator {}
+// logical
+export class Or extends BinaryOperator {}
+Or.prototype.type = Boolean;
+export class And extends BinaryOperator {}
+And.prototype.type = Boolean;
+export class Equal extends BinaryOperator {
+    fold() {
+        if (this.isConstant) {
+            let lhs = this.lhs.fold();
+            let rhs = this.rhs.fold();
+            if (lhs instanceof Primitive && rhs instanceof Primitive) {
+                return this.replace(new Boolean(lhs.value === rhs.value));
+            }
+        }
+        return this;
+    }
+}
+Equal.prototype.type = Boolean;
+export class NotEqual extends BinaryOperator {
+    fold() {
+        if (this.isConstant) {
+            let lhs = this.lhs.fold();
+            let rhs = this.rhs.fold();
+            if (lhs instanceof Primitive && rhs instanceof Primitive) {
+                return this.replace(new Boolean(lhs.value !== rhs.value));
+            }
+        }
+        return this;
+    }
+}
+NotEqual.prototype.type = Boolean;
+export class GreaterThan extends BinaryOperator {
+    fold() {
+        if (this.isConstant) {
+            let lhs = this.lhs.fold();
+            let rhs = this.rhs.fold();
+            if (lhs instanceof Number && rhs instanceof Number)
+                return this.replace(new Boolean(lhs.value > rhs.value));
+        }
+        return this;
+    }
+}
+GreaterThan.prototype.type = Boolean;
+export class LessThan extends BinaryOperator {
+    fold() {
+        if (this.isConstant) {
+            let lhs = this.lhs.fold();
+            let rhs = this.rhs.fold();
+            if (lhs instanceof Number && rhs instanceof Number)
+                return this.replace(new Boolean(lhs.value < rhs.value));
+        }
+        return this;
+    }
+}
+LessThan.prototype.type = Boolean;
+export class GreaterThanEqual extends BinaryOperator {
+    fold() {
+        if (this.isConstant) {
+            let lhs = this.lhs.fold();
+            let rhs = this.rhs.fold();
+            if (lhs instanceof Number && rhs instanceof Number)
+                return this.replace(new Boolean(lhs.value >= rhs.value));
+        }
+        return this;
+    }
+}
+GreaterThanEqual.prototype.type = Boolean;
+export class LessThanEqual extends BinaryOperator {
+    fold() {
+        if (this.isConstant) {
+            let lhs = this.lhs.fold();
+            let rhs = this.rhs.fold();
+            if (lhs instanceof Number && rhs instanceof Number)
+                return this.replace(new Boolean(lhs.value <= rhs.value));
+        }
+        return this;
+    }
+}
+LessThanEqual.prototype.type = Boolean;
+// arithmetical
+export class Plus extends BinaryOperator {
+    fold() {
+        if (this.isConstant) {
+            let lhs = this.lhs.fold();
+            let rhs = this.rhs.fold();
+            if (lhs instanceof String && rhs instanceof String)
+                return this.replace(new String(lhs.value + rhs.value));
+            if (lhs instanceof Integer && rhs instanceof Integer)
+                return this.replace(new Integer(lhs.value + rhs.value));
+            if (lhs instanceof Number && rhs instanceof Number)
+                return this.replace(new Float(lhs.value + rhs.value));
+        }
+        return this;
+    }
+}
+export class Minus extends BinaryOperator {
+    fold() {
+        if (this.isConstant) {
+            let lhs = this.lhs.fold();
+            let rhs = this.rhs.fold();
+            if (lhs instanceof Integer && rhs instanceof Integer)
+                return this.replace(new Integer(lhs.value - rhs.value));
+            if (lhs instanceof Number && rhs instanceof Number)
+                return this.replace(new Float(lhs.value - rhs.value));
+        }
+        return this;
+    }
+}
+export class Multiply extends BinaryOperator {
+    fold() {
+        if (this.isConstant) {
+            let lhs = this.lhs.fold();
+            let rhs = this.rhs.fold();
+            if (lhs instanceof Integer && rhs instanceof Integer)
+                return this.replace(new Integer(lhs.value * rhs.value));
+            if (lhs instanceof Number && rhs instanceof Number)
+                return this.replace(new Float(lhs.value * rhs.value));
+        }
+        return this;
+    }
+}
+export class Divide extends BinaryOperator {
+    fold() {
+        if (this.isConstant) {
+            let lhs = this.lhs.fold();
+            let rhs = this.rhs.fold();
+            if (lhs instanceof Number && rhs instanceof Number)
+                return this.replace(new Float(lhs.value / rhs.value));
+        }
+        return this;
+    }
+}
+Divide.prototype.type = Float;
+export class Modular extends BinaryOperator {
+    fold() {
+        if (this.isConstant) {
+            let lhs = this.lhs.fold();
+            let rhs = this.rhs.fold();
+            if (lhs instanceof Integer && rhs instanceof Integer)
+                return this.replace(new Integer(lhs.value % rhs.value));
+            if (lhs instanceof Number && rhs instanceof Number)
+                return this.replace(new Float(lhs.value % rhs.value));
+        }
+        return this;
+    }
+}
